fix: Java enum mapping is broken
Fixes: #1199
