+++ ui/src/app/lib/atlasmap-data-mapper/models/mapping-definition.model.ts
@@ -14,13 +14,10 @@
     limitations under the License.
 */
 
-import { MappingModel, MappedField } from './mapping.model';
+import { MappingModel } from './mapping.model';
 import { LookupTable } from '../models/lookup-table.model';
-import { ConfigModel } from '../models/config.model';
 import { Field } from '../models/field.model';
-import { FieldAction, FieldActionArgumentValue, FieldActionDefinition, Multiplicity } from './field-action.model';
-import { TransitionMode } from './transition.model';
-import { DocumentDefinition, PaddingField } from '../models/document-definition.model';
+import { DocumentDefinition } from '../models/document-definition.model';
 
 import { DataMapperUtil } from '../common/data-mapper-util';
 
@@ -53,35 +50,6 @@ export class MappingDefinition {
     return this.tablesByName[name];
   }
 
-  detectTableIdentifiers() {
-    for (const t of this.getTables()) {
-      if (t.sourceIdentifier && t.targetIdentifier) {
-        continue;
-      }
-      let tableChanged = false;
-      const m: MappingModel = this.getFirstMappingForLookupTable(t.name);
-      if (m != null && m.transition.lookupTableName != null) {
-        if (!t.sourceIdentifier) {
-          const inputField: Field = m.getFields(true)[0];
-          if (inputField) {
-            t.sourceIdentifier = inputField.classIdentifier;
-            tableChanged = true;
-          }
-        }
-        if (!t.targetIdentifier) {
-          const outputField: Field = m.getFields(false)[0];
-          if (outputField) {
-            t.targetIdentifier = outputField.classIdentifier;
-            tableChanged = true;
-          }
-        }
-      }
-    }
-    for (const m of this.mappings) {
-      this.initializeMappingLookupTable(m);
-    }
-  }
-
   getTableBySourceTarget(sourceIdentifier: string, targetIdentifier: string): LookupTable {
     const key: string = sourceIdentifier + ':' + targetIdentifier;
     return this.tablesBySourceTargetKey[key];
@@ -108,139 +76,6 @@ export class MappingDefinition {
         delete this.tablesByName[name];
         delete this.tablesBySourceTargetKey[iokey];
       }
-    }
-  }
-
-  getFirstMappingForLookupTable(lookupTableName: string): MappingModel {
-    for (const m of this.mappings) {
-      if (m.transition.lookupTableName === lookupTableName) {
-        return m;
-      }
-    }
-    return null;
-  }
-
-  removeStaleMappings(cfg: ConfigModel): void {
-    let index = 0;
-    let sourceFieldPaths: string[] = [];
-    for (const doc of cfg.getDocs(true)) {
-      sourceFieldPaths = sourceFieldPaths.concat(Field.getFieldPaths(doc.getAllFields()));
-    }
-    let targetSourcePaths: string[] = [];
-    for (const doc of cfg.getDocs(false)) {
-      targetSourcePaths = targetSourcePaths.concat(Field.getFieldPaths(doc.getAllFields()));
-    }
-    while (index < this.mappings.length) {
-      const mapping: MappingModel = this.mappings[index];
-      const mappingIsStale: boolean = this.isMappingStale(mapping, sourceFieldPaths, targetSourcePaths);
-      if (mappingIsStale) {
-        this.mappings.splice(index, 1);
-      } else {
-        index++;
-      }
-    }
-  }
-
-  isMappingStale(mapping: MappingModel, sourceFieldPaths: string[], targetSourcePaths: string[]): boolean {
-
-    for (const field of mapping.getFields(true)) {
-      if (!(field instanceof PaddingField) && sourceFieldPaths.indexOf(field.path) === -1) {
-        return true;
-      }
-    }
-    for (const field of mapping.getFields(false)) {
-      if (!(field instanceof PaddingField) && targetSourcePaths.indexOf(field.path) === -1) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  initializeMappingLookupTable(m: MappingModel): void {
-    if (!(m.transition.mode === TransitionMode.ENUM
-      && m.transition.lookupTableName == null
-      && m.getFields(true).length === 1
-      && m.getFields(false).length === 1)) {
-      return;
-    }
-    let inputClassIdentifier: string = null;
-    let outputClassIdentifier: string = null;
-
-    const inputField: Field = m.getFields(true)[0];
-    if (inputField) {
-      inputClassIdentifier = inputField.classIdentifier;
-    }
-    const outputField: Field = m.getFields(true)[0];
-    if (outputField) {
-      outputClassIdentifier = outputField.classIdentifier;
-    }
-    if (inputClassIdentifier && outputClassIdentifier) {
-      let table: LookupTable = this.getTableBySourceTarget(inputClassIdentifier, outputClassIdentifier);
-      if (table == null) {
-        table = new LookupTable();
-        table.sourceIdentifier = inputClassIdentifier;
-        table.targetIdentifier = outputClassIdentifier;
-        this.addTable(table);
-        m.transition.lookupTableName = table.name;
-      } else {
-        m.transition.lookupTableName = table.name;
-      }
-    }
-  }
-
-  /**
-   * Return a document map for either the sources or targets panel contents.
-   *
-   * @param cfg
-   * @param isSource
-   */
-  private getDocMap(cfg: ConfigModel, isSource: boolean): any {
-    const docMap: any = {};
-    for (const doc of cfg.getDocs(isSource)) {
-      docMap[doc.uri] = doc;
-    }
-    return docMap;
-  }
-
-  updateMappingsFromDocuments(cfg: ConfigModel): void {
-    const sourceDocMap: any = this.getDocMap(cfg, true);
-    const targetDocMap: any = this.getDocMap(cfg, false);
-
-    for (const mapping of this.mappings) {
-      this.updateMappedFieldsFromDocuments(mapping, cfg, sourceDocMap, true);
-      this.updateMappedFieldsFromDocuments(mapping, cfg, targetDocMap, false);
-    }
-    for (const doc of cfg.getAllDocs()) {
-      if (doc.id == null) {
-        doc.id = 'DOC.' + doc.name + '.' + Math.floor((Math.random() * 1000000) + 1).toString();
-      }
-    }
-  }
-
-  updateDocumentNamespacesFromMappings(cfg: ConfigModel): void {
-    const docs: DocumentDefinition[] = cfg.getDocs(false);
-
-    for (const parsedDoc of this.parsedDocs) {
-      if (!parsedDoc) {
-        continue;
-      }
-      if (parsedDoc.isSource) {
-        continue;
-      }
-      if (parsedDoc.namespaces.length === 0) {
-        continue;
-      }
-
-      const doc = DocumentDefinition.getDocumentByIdentifier(parsedDoc.id, docs);
-      if (doc == null) {
-        cfg.errorService.error('Could not find document with identifier \'' + parsedDoc.id
-          + '\' for namespace override.',
-          { 'identifier': parsedDoc.id, 'parsedDoc': parsedDoc, 'docs': docs });
-        continue;
-      }
-
-      doc.namespaces = [].concat(parsedDoc.namespaces);
     }
   }
 
@@ -274,147 +109,4 @@ export class MappingDefinition {
     return DataMapperUtil.removeItemFromArray(m, this.mappings);
   }
 
-  removeFieldFromAllMappings(field: Field): void {
-    for (const mapping of this.getAllMappings(true)) {
-      const mappedField: MappedField = mapping.getMappedFieldForField(field);
-      if (mappedField != null) {
-        mapping.removeMappedField(mappedField);
-        if (mapping.isEmpty()) {
-          this.removeMapping(mapping);
-        }
-      }
-    }
-  }
-
-  /**
-   * Remove any mappings referencing the specified document ID.
-   *
-   * @param docId - Specified document ID
-   * @param cfg
-   */
-  removeDocumentReferenceFromAllMappings(docId: string, cfg: ConfigModel) {
-    for (const mapping of this.getAllMappings(true)) {
-      for (const mappedField of mapping.getAllFields()) {
-        if (!(mappedField instanceof PaddingField) && (mappedField.docDef.id === docId)) {
-          this.removeFieldFromAllMappings(mappedField);
-          this.removeMapping(mapping);
-          if (mapping === cfg.mappings.activeMapping) {
-            cfg.mappingService.deselectMapping();
-          }
-        }
-      }
-    }
-  }
-
-  updateMappedFieldsFromDocuments(mapping: MappingModel, cfg: ConfigModel, docMap: any, isSource: boolean): void {
-    let mappedFields: MappedField[] = mapping.getMappedFields(isSource);
-    let mappedFieldIndex = -1;
-
-    for (const mappedField of mappedFields) {
-      let doc: DocumentDefinition = null;
-      mappedFieldIndex += 1;
-
-      if (mappedField.parsedData.fieldIsProperty) {
-        doc = cfg.propertyDoc;
-      } else if (mappedField.parsedData.fieldIsConstant) {
-        doc = cfg.constantDoc;
-      } else {
-        if (docMap === null) {
-          docMap = this.getDocMap(cfg, isSource);
-        }
-        doc = docMap[mappedField.parsedData.parsedDocURI] as DocumentDefinition;
-        if (doc == null) {
-          if (mappedField.parsedData.parsedName != null) {
-            cfg.errorService.error('Could not find document for mapped field \'' + mappedField.parsedData.parsedName +
-              '\' at URI ' + mappedField.parsedData.parsedDocURI, null);
-          }
-          continue;
-        }
-
-        if (mappedField.parsedData.parsedDocID == null) {
-          cfg.errorService.error('Could not find doc ID for mapped field ' + mappedField.parsedData.parsedName, null);
-          continue;
-        }
-        doc.id = mappedField.parsedData.parsedDocID;
-      }
-      mappedField.field = null;
-      if (!mappedField.parsedData.userCreated) {
-        mappedField.field = doc.getField(mappedField.parsedData.parsedPath);
-      }
-      if (mappedField.field == null) {
-        if (mappedField.parsedData.fieldIsConstant || mappedField.parsedData.fieldIsProperty) {
-          const constantField: Field = new Field();
-          constantField.value = mappedField.parsedData.parsedValue;
-          constantField.type = mappedField.parsedData.parsedValueType;
-          constantField.displayName = constantField.value;
-          constantField.name = constantField.value;
-          constantField.path = constantField.value;
-          constantField.userCreated = true;
-          mappedField.field = constantField;
-          doc.addField(constantField);
-        } else if (mappedField.parsedData.userCreated || mappedField.parsedData.parsedPath) {
-          const path: string = mappedField.parsedData.parsedPath;
-
-          mappedField.field = new Field();
-          mappedField.field.serviceObject.jsonType = 'io.atlasmap.xml.v2.XmlField';
-          mappedField.field.path = path;
-          mappedField.field.type = mappedField.parsedData.parsedValueType;
-          mappedField.field.userCreated = true;
-
-          const lastSeparator: number = path.lastIndexOf('/');
-
-          const parentPath: string = (lastSeparator > 0) ? path.substring(0, lastSeparator) : null;
-          let fieldName: string = (lastSeparator === -1) ? path : path.substring(lastSeparator + 1);
-          let namespaceAlias: string = null;
-          if (fieldName.indexOf(':') !== -1) {
-            namespaceAlias = fieldName.split(':')[0];
-            fieldName = fieldName.split(':')[1];
-          }
-
-          mappedField.field.name = fieldName;
-          mappedField.field.displayName = fieldName;
-          mappedField.field.isAttribute = (fieldName.indexOf('@') !== -1);
-          mappedField.field.namespaceAlias = namespaceAlias;
-
-          if (parentPath != null) {
-            mappedField.field.parentField = doc.getField(parentPath);
-          }
-
-          doc.addField(mappedField.field);
-        } else {
-          cfg.errorService.error('Could not find field from document for mapped field \'' + mappedField.parsedData.parsedName + '\'',
-            { 'mappedField': mappedField, 'doc': doc });
-          return;
-        }
-      }
-
-      // Process field actions.
-      mappedField.actions = [];
-      if (mappedField.parsedData.parsedActions.length > 0) {
-
-        for (const action of mappedField.parsedData.parsedActions) {
-          let actionDefinition = null;
-          if (action.name === 'CustomAction') {
-            actionDefinition = cfg.fieldActionService.getActionDefinitionForName(action.argumentValues[0].value, Multiplicity.ONE_TO_ONE);
-          } else {
-            actionDefinition = cfg.fieldActionService.getActionDefinitionForName(action.name, Multiplicity.ONE_TO_ONE);
-          }
-          if (actionDefinition == null) {
-            cfg.errorService.error('Could not find field action definition for action \'' + action.name + '\'', null);
-            continue;
-          }
-          actionDefinition.populateFieldAction(action);
-          mappedField.actions.push(action);
-        }
-      }
-
-      const zeroBasedIndex = +mappedField.parsedData.parsedIndex;
-      mappedFields = mapping.getMappedFields(isSource);
-      if (zeroBasedIndex <= mappedFieldIndex) {
-        mappedFields[zeroBasedIndex] = mappedField;
-      } else {
-        cfg.mappingService.addPlaceholders(zeroBasedIndex - mappedFieldIndex, mapping, mappedFieldIndex, isSource);
-      }
-    }
-  }
 }
